[#1] Split project into individual files
