+++ src/main/java/com/garciat/typeclasses/api/hkt/TApp.java
@@ -1,0 +1,11 @@
+package com.garciat.typeclasses.api.hkt;
+
+import com.garciat.typeclasses.api.hkt.Kind.KArr;
+import com.garciat.typeclasses.api.hkt.Kind.KStar;
+
+/**
+ * Full application of a unary type constructor.
+ *
+ * <p>TApp :: (* -> *) -> * -> *
+ */
+public interface TApp<Tag extends Kind<KArr<KStar>>, A> extends Kind<KStar> {}
