+++ src/main/java/com/garciat/typeclasses/api/hkt/TagBase.java
@@ -1,0 +1,3 @@
+package com.garciat.typeclasses.api.hkt;
+
+public abstract class TagBase<K extends Kind.Base> implements Kind<K> {}
