+++ src/main/java/com/garciat/typeclasses/classes/Ordering.java
@@ -1,0 +1,7 @@
+package com.garciat.typeclasses.classes;
+
+public enum Ordering {
+  LT,
+  EQ,
+  GT
+}
