+++ src/main/java/com/garciat/typeclasses/types/F1.java
@@ -1,0 +1,12 @@
+package com.garciat.typeclasses.types;
+
+import java.util.function.Function;
+
+@FunctionalInterface
+public interface F1<A, R> {
+  R apply(A a);
+
+  static <A, R> F1<A, R> of(Function<A, R> f) {
+    return f::apply;
+  }
+}
