--- src/test/java/com/garciat/typeclasses/MainTest.java
@@ -1,12 +1,0 @@
-package com.garciat.typeclasses;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-final class MainTest {
-  @Test
-  void example() {
-    assertTrue(true);
-  }
-}
